requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Pillow==7.0.0
Markdown==3.3.7